+++ CHANGELOG.md
@@ -1,10 +1,9 @@
 ## HEAD (Unreleased)
-_(none)_
+* Upgrade to v2.9.0 of the Docker Terraform Provider
 
 ---
 
 ## 2.5.0 (2020-12-07)
-
 * Upgrade to v2.8.0 of the Docker Terraform Provider
 
 ## 2.4.1 (2020-11-24)